backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
pdfplumber==0.11.0
Pillow>=10.4.0
pytesseract==0.3.13
werkzeug==2.3.8
opencv-python>=4.10.0
numpy>=1.26.4
spacy>=3.7.0
